requirements.txt
openai-whisper
pydub
numpy